Implementation of the json:api change
The response of the http request is now being
parsed to dissolve the response body object. This
means that only the actual data will be passed
through and in case the key 'data' is set to null,
an exception is thrown which can be caught in
order to redirect the user to the 404 page.